api/requirements.txt
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

# Google APIs
google-api-python-client==2.116.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# GitHub API
PyGithub==2.1.1

# Scraping
gallery-dl==1.26.8
yt-dlp>=2025.1.15
beautifulsoup4==4.12.3
httpx==0.26.0

# Email
python-multipart==0.0.6
aiosmtplib==3.0.1
email-validator==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utils
python-dateutil==2.8.2
pytz==2023.3
aiofiles==23.2.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
